KAR-1935 Fix data types for spfrx

+++ tests/acceptance/test_noise_diode_attributes.py
@@ -24,7 +24,7 @@ def reset_diode_params(dish_manager_proxy: tango.DeviceProxy):
     [
         ("noiseDiodeMode", NoiseDiodeMode.PERIODIC),
         ("periodicnoisediodepars", np.array([1, 2, 3], dtype=np.uint32)),
-        ("pseudorandomnoisediodepars", np.array([1, 2, 3], dtype=np.int64)),
+        ("pseudorandomnoisediodepars", np.array([1, 2, 3], dtype=np.uint32)),
     ],
 )
 def test_set_noise_diode_attribute(
